Add support for hashing unique part of subject-id
