+++ modules/saml/src/Auth/Process/PairwiseID.php
@@ -4,10 +4,8 @@ declare(strict_types=1);
 
 namespace SimpleSAML\Module\saml\Auth\Process;
 
-use SimpleSAML\{Auth, Utils};
 use SimpleSAML\SAML2\Constants as C;
 
-use function hash_hmac;
 use function strtolower;
 
 /**
@@ -44,11 +42,6 @@ class PairwiseID extends SubjectID
      */
     public const NAME = 'PairwiseID';
 
-    /**
-     * @var \SimpleSAML\Utils\Config
-     */
-    protected Utils\Config $configUtils;
-
 
     /**
      * Initialize this filter.
@@ -59,8 +52,6 @@ class PairwiseID extends SubjectID
     public function __construct(array &$config, $reserved)
     {
         parent::__construct($config, $reserved);
-
-        $this->configUtils = new Utils\Config();
     }
 
 
@@ -87,23 +78,11 @@ class PairwiseID extends SubjectID
         }
 
         // Calculate hash
-        $salt = $this->configUtils->getSecretSalt();
-        $hash = hash_hmac('sha256', $userID . '|' . $sp_entityid, $salt, false);
+        $hash = $this->calculateHash($userID . '|' . $sp_entityid);
 
-        $value = $hash . '@' . strtolower($scope);
+        $value = strtolower($hash . '@' . $scope);
         $this->validateGeneratedIdentifier($value);
 
         $state['Attributes'][C::ATTR_PAIRWISE_ID] = [$value];
-    }
-
-
-    /**
-     * Inject the \SimpleSAML\Utils\Config dependency.
-     *
-     * @param \SimpleSAML\Utils\Config $configUtils
-     */
-    public function setConfigUtils(Utils\Config $configUtils): void
-    {
-        $this->configUtils = $configUtils;
     }
 }
